Set experimental headers bu default on Agent administration and responses client.
