+++ sdk/ai/Azure.AI.Extensions.OpenAI/assets.json
@@ -1,7 +1,7 @@
 {
   "AssetsFileName": "assets.json",
   "AssetsRepo": "Azure/azure-sdk-assets",
-  "Tag": "net/ai/Azure.AI.Extensions.OpenAI_b277d70189",
+  "Tag": "net/ai/Azure.AI.Extensions.OpenAI_ec0d47d345",
   "AssetsRepoPrefixPath": "net",
   "TagPrefix": "net/ai/Azure.AI.Extensions.OpenAI"
 }
