+++ sdk/ai/Azure.AI.Extensions.OpenAI/tests/Samples/Sample_CodeAgentReminderTool.cs
@@ -2,8 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.ClientModel.Primitives;
-using System.Collections.Generic;
 using System.IO;
 using System.Runtime.CompilerServices;
 using System.Threading;
