+++ sdk/ai/Azure.AI.Projects.Agents/src/Custom/Internal/FoundryFeaturesPolicy.cs
@@ -1,0 +1,26 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.ClientModel.Primitives;
+using System.Collections.Generic;
+using System.Threading.Tasks;
+
+namespace Azure.AI.Projects.Agents.Custom.Internal
+{
+    internal class FoundryFeaturesPolicy(string feature) : PipelinePolicy
+    {
+        private const string _FEATURE_HEADER = "Foundry-Features";
+
+        public override void Process(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
+        {
+            message.Request.Headers.Set(_FEATURE_HEADER, feature);
+            ProcessNext(message, pipeline, currentIndex);
+        }
+
+        public override async ValueTask ProcessAsync(PipelineMessage message, IReadOnlyList<PipelinePolicy> pipeline, int currentIndex)
+        {
+            message.Request.Headers.Set(_FEATURE_HEADER, feature);
+            await ProcessNextAsync(message, pipeline, currentIndex).ConfigureAwait(false);
+        }
+    }
+}
