+++ sdk/ai/Azure.AI.Projects.Agents/tests/AgentsTestBase.cs
@@ -31,9 +31,6 @@ public class AgentsTestBase : RecordedTestBase<AgentsTestEnvironment>
     protected readonly string MEMORY_STORE_SCOPE = "user_123";
     protected readonly int PAGE_SIZE = 3;
 
-    protected const string FOUNDRY_HEADER = "Foundry-Features";
-    protected const string FOUNDRY_HEADER_VALUE = "MemoryStores=V1Preview,ContainerAgents=V1Preview,WorkflowAgents=V1Preview,Evaluations=V1Preview,Schedules=V1Preview,RedTeams=V1Preview,AgentEndpoints=V1Preview,Skills=V1Preview,Insights=V1Preview,DataGenerationJobs=V1Preview,Models=V1Preview,AgentsOptimization=V2Preview,Routines=V1Preview,ExternalAgents=V1Preview,DraftAgents=V1Preview";
-
     public AgentsTestBase(bool isAsync, RecordedTestMode? testMode = null) : base(isAsync, testMode)
     {
         // Please note that in System.ClientModel, the recording mode is taken from CLIENTMODEL_TEST_MODE
@@ -140,7 +137,6 @@ public class AgentsTestBase : RecordedTestBase<AgentsTestEnvironment>
                 {
                     message.NetworkTimeout = TimeSpan.FromMinutes(5);
                 }
-                message.Request.Headers.Set(FOUNDRY_HEADER, FOUNDRY_HEADER_VALUE);
             }),
             PipelinePosition.PerCall);
         return CreateProxyFromClient(new AgentAdministrationClient(new(TestEnvironment.FOUNDRY_PROJECT_ENDPOINT), GetTestTokenProvider(), InstrumentClientOptions(options)));
@@ -159,7 +155,6 @@ public class AgentsTestBase : RecordedTestBase<AgentsTestEnvironment>
 
     public static void AssertListEqual(string[] expected, List<string> observed)
     {
-        // Assert.AreEqual(expected.Length, observed.Count, $"The length of arrays are different. Expected: {expected}, Observed: {observed.ToArray()}");
         HashSet<string> expectedHash = [.. expected];
         HashSet<string> observedHash = [.. observed];
         if (!expectedHash.SetEquals(observedHash))
@@ -363,12 +358,7 @@ public class AgentsTestBase : RecordedTestBase<AgentsTestEnvironment>
     {
         if (Mode == RecordedTestMode.Playback)
             return;
-        AgentAdministrationClientOptions options = new();
-        options.AddPolicy(
-            new HeaderTestPolicy(new Dictionary<string, string>()
-            { {FOUNDRY_HEADER, FOUNDRY_HEADER_VALUE} }),
-            PipelinePosition.PerCall);
-        AgentAdministrationClient agentsClient = new(new(TestEnvironment.FOUNDRY_PROJECT_ENDPOINT), GetTestTokenProvider(), options);
+        AgentAdministrationClient agentsClient = new(new(TestEnvironment.FOUNDRY_PROJECT_ENDPOINT), GetTestTokenProvider());
 
         // Remove Agents.
         foreach (ProjectsAgentVersion ag in agentsClient.GetAgentVersions(agentName: AGENT_NAME))
