+++ sdk/ai/Azure.AI.Projects.Agents/tests/Samples/Sample_AgentsOptimizationCRUD.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT License.
 
 using System;
-using System.ClientModel.Primitives;
 using System.Collections.Generic;
 using System.Text.Json;
 using System.Threading;
@@ -88,9 +87,7 @@ public class Sample_AgentsOptimizationCRUD : SamplesBase
         var modelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME;
         var anotherModelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME2;
 #endif
-        AgentAdministrationClientOptions options = new();
-        options.AddPolicy(new FeaturePolicy("AgentsOptimization=V2Preview"), PipelinePosition.PerCall);
-        AgentAdministrationClient agentsClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential(), options: options);
+        AgentAdministrationClient agentsClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential());
         AgentOptimizationJobs jobsClient = agentsClient.GetAgentOptimizationJobs();
         #endregion
 
@@ -197,9 +194,7 @@ public class Sample_AgentsOptimizationCRUD : SamplesBase
         var modelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME;
         var anotherModelDeploymentName = TestEnvironment.FOUNDRY_MODEL_NAME2;
 #endif
-        AgentAdministrationClientOptions options = new();
-        options.AddPolicy(new FeaturePolicy("AgentsOptimization=V2Preview"), PipelinePosition.PerCall);
-        AgentAdministrationClient agentsClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential(), options: options);
+        AgentAdministrationClient agentsClient = new(endpoint: new Uri(projectEndpoint), tokenProvider: new DefaultAzureCredential());
         AgentOptimizationJobs jobsClient = agentsClient.GetAgentOptimizationJobs();
 
         #region Snippet:Sample_CreateAgent_AgentsOptimization_Sync
