+++ sdk/ai/Azure.AI.Projects.Agents/tests/Samples/Sample_SessionFiles.cs
@@ -3,7 +3,6 @@
 
 using System;
 using System.ClientModel;
-using System.ClientModel.Primitives;
 using System.IO;
 using System.Threading;
 using System.Threading.Tasks;
