+++ sdk/ai/Azure.AI.Projects/assets.json
@@ -2,5 +2,5 @@
   "AssetsRepo": "Azure/azure-sdk-assets",
   "AssetsRepoPrefixPath": "net",
   "TagPrefix": "net/ai/Azure.AI.Projects",
-  "Tag": "net/ai/Azure.AI.Projects_3194e73d0b"
+  "Tag": "net/ai/Azure.AI.Projects_63626cf80e"
 }
